Fixed Marker settings and Camera Position (Android)
When creating the map, you can pass the option (centerPin=true) to create a fixed (fake) marker in the exact center of the map.
It will float over the map and it's bottom-center is aligned with the center of the map.
Image used as a marker and size can be passed as options in the centerPinOptions key.
The marker can be created after the map is created with the createCenterPin method.
Also, a getCameraPosition method has been added, which returns the Lat and Lng of the Camera (=center of the map).
For now, these features are Android only.